+++ src/main/java/com/example/happy_fitness/controller/ContactController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -26,7 +25,7 @@ public class ContactController {
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Facility> getBlogById(@PathVariable("id") BigInteger coachId){
+	public ResponseEntity<Facility> getBlogById(@PathVariable("id") Long coachId){
 		return new ResponseEntity<Facility>(facilityService.findById(coachId), HttpStatus.OK);
 	}
 }
