+++ src/main/java/com/example/happy_fitness/controller/FacilityProductController.java
@@ -13,8 +13,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
-
 @RestController
 @RequestMapping("/api/facility-product")
 @Slf4j
@@ -25,8 +23,8 @@ public class FacilityProductController {
 
     @PostMapping("/update")
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_MANAGER')")
-    public ResponseEntity<BaseResponse<String>> order(@RequestParam BigInteger productId,
-                                                      @RequestParam BigInteger facilityId,
+    public ResponseEntity<BaseResponse<String>> order(@RequestParam Long productId,
+                                                      @RequestParam Long facilityId,
                                                       @RequestBody FacilityProduct facilityProduct,
                                                       @AuthenticationPrincipal UserDetails userDetails) {
         try {
