+++ src/main/java/com/example/happy_fitness/controller/ProductController.java
@@ -17,7 +17,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.List;
 
@@ -34,10 +33,10 @@ public class ProductController {
 
     @GetMapping("")
     public ResponseEntity<BaseResponse<List<ProductDto>>> getProduct(
-            @RequestParam BigInteger facilityId,
+            @RequestParam Long facilityId,
             @RequestParam(required = false) String status,
-            @RequestParam(required = false) BigInteger categoryId,
-            @RequestParam(required = false) BigInteger supplierId,
+            @RequestParam(required = false) Long categoryId,
+            @RequestParam(required = false) Long supplierId,
             @RequestParam(required = false) Float minPrice,
             @RequestParam(required = false) Float maxPrice
     ) {
@@ -50,7 +49,7 @@ public class ProductController {
     }
 
     @GetMapping("/{code}")
-    public ResponseEntity<BaseResponse<List<ProductDto>>> getProductDetail(@PathVariable String code, @RequestParam BigInteger facilityId) {
+    public ResponseEntity<BaseResponse<List<ProductDto>>> getProductDetail(@PathVariable String code, @RequestParam Long facilityId) {
         try {
             return ResponseEntity.ok(BaseResponse.ok(productService.findProductDetail(facilityId, code)));
         } catch (Exception e) {
@@ -79,7 +78,7 @@ public class ProductController {
 //            @RequestParam("product") String productStr,
 //            @RequestParam(name = "image", required = false) MultipartFile image,
             @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable BigInteger id) {
+            @PathVariable Long id) {
         try {
 //            return ResponseEntity.ok(BaseResponse.ok(productService.updateCustom(objectMapper.readValue(productStr, Product.class), id, image)));
             return ResponseEntity.ok(BaseResponse.ok(productService.update(product, id, userDetails)));
@@ -91,7 +90,7 @@ public class ProductController {
 
     @PostMapping("/delete/{id}")
     @PreAuthorize("hasAuthority('ROLE_ADMIN')")
-    public ResponseEntity<BaseResponse<String>> deleteProduct(@PathVariable BigInteger id) {
+    public ResponseEntity<BaseResponse<String>> deleteProduct(@PathVariable Long id) {
         try {
             productService.delete(Arrays.asList(id));
             return ResponseEntity.ok(BaseResponse.ok(HttpStatus.OK.getReasonPhrase()));
@@ -103,7 +102,7 @@ public class ProductController {
 
     @PostMapping("/active/{id}")
     @PreAuthorize("hasAuthority('ROLE_ADMIN')")
-    public ResponseEntity<BaseResponse<String>> active(@PathVariable BigInteger id) {
+    public ResponseEntity<BaseResponse<String>> active(@PathVariable Long id) {
         try {
             productService.active(id);
             return ResponseEntity.ok(BaseResponse.ok(HttpStatus.OK.getReasonPhrase()));
