+++ src/main/java/com/example/happy_fitness/controller/TicketController.java
@@ -13,7 +13,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @RestController
@@ -37,7 +36,7 @@ public class TicketController {
     }
 
     @GetMapping("")
-    public ResponseEntity<BaseResponse<List<Ticket>>> findTickets(@RequestParam BigInteger facilityId) {
+    public ResponseEntity<BaseResponse<List<Ticket>>> findTickets(@RequestParam Long facilityId) {
         try {
             return ResponseEntity.ok(BaseResponse.ok(ticketService.findAllByFacilityId(facilityId)));
         } catch (Exception e) {
@@ -47,7 +46,7 @@ public class TicketController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<BaseResponse<List<Ticket>>> findTicketDetail(@PathVariable BigInteger id) {
+    public ResponseEntity<BaseResponse<List<Ticket>>> findTicketDetail(@PathVariable Long id) {
         try {
             return ResponseEntity.ok(BaseResponse.ok(ticketService.findTicketDetail(id)));
         } catch (Exception e) {
@@ -58,7 +57,7 @@ public class TicketController {
 
     @PostMapping("/deactivate/{id}")
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_MANAGER')")
-    public ResponseEntity<BaseResponse<List<Ticket>>> findTicketDetail(@PathVariable BigInteger id,
+    public ResponseEntity<BaseResponse<List<Ticket>>> findTicketDetail(@PathVariable Long id,
                                                                        @AuthenticationPrincipal UserDetails userDetails) {
         try {
             return ResponseEntity.ok(BaseResponse.ok(ticketService.deactivate(userDetails, id)));
@@ -70,7 +69,7 @@ public class TicketController {
 
     @PostMapping("/update/{id}")
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN', 'ROLE_MANAGER')")
-    public ResponseEntity<BaseResponse<String>> update(@PathVariable BigInteger id, @RequestBody Ticket ticket, @AuthenticationPrincipal UserDetails userDetails) {
+    public ResponseEntity<BaseResponse<String>> update(@PathVariable Long id, @RequestBody Ticket ticket, @AuthenticationPrincipal UserDetails userDetails) {
         try {
             return ResponseEntity.ok(BaseResponse.ok(ticketService.update(ticket, id, userDetails)));
         } catch (Exception e) {
