+++ src/main/java/com/example/happy_fitness/dto/CartDto.java
@@ -2,12 +2,10 @@ package com.example.happy_fitness.dto;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class CartDto {
-    private BigInteger id;
-    private BigInteger facilityProductId;
+    private Long id;
+    private Long facilityProductId;
     private String code;
     private String name;
     private String image;
