+++ src/main/java/com/example/happy_fitness/dto/UserDto.java
@@ -2,12 +2,11 @@ package com.example.happy_fitness.dto;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 @Data
 public class UserDto {
-    private BigInteger id;
+    private Long id;
     private String username;
     private String fullName;
     private String email;
@@ -18,7 +17,7 @@ public class UserDto {
     private String facility;
     private String roleName;
 
-    public UserDto(BigInteger id, String username, String fullName, String email, Boolean gender, Date dob, String phoneNumber, String address, String facility, String roleName) {
+    public UserDto(Long id, String username, String fullName, String email, Boolean gender, Date dob, String phoneNumber, String address, String facility, String roleName) {
         this.id = id;
         this.username = username;
         this.fullName = fullName;
