+++ src/main/java/com/example/happy_fitness/repository/CustomerTicketRepository.java
@@ -4,14 +4,13 @@ import com.example.happy_fitness.entity.CustomerTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface CustomerTicketRepository extends JpaRepository<CustomerTicket, BigInteger> {
+public interface CustomerTicketRepository extends JpaRepository<CustomerTicket, Long> {
     List<CustomerTicket> findAllByStatusIsTrue();
 
-    List<CustomerTicket> findAllByCustomer_IdAndStatusIsTrue(BigInteger customerId);
+    List<CustomerTicket> findAllByCustomer_IdAndStatusIsTrue(Long customerId);
 
     List<CustomerTicket> findAllByCustomer_UsernameOrderByUpdatedDateDesc(String username);
 }
