+++ src/main/java/com/example/happy_fitness/repository/FacilityRepository.java
@@ -3,7 +3,5 @@ package com.example.happy_fitness.repository;
 import com.example.happy_fitness.entity.Facility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface FacilityRepository extends JpaRepository<Facility, BigInteger> {
+public interface FacilityRepository extends JpaRepository<Facility, Long> {
 }
