+++ src/main/java/com/example/happy_fitness/repository/MailTemplateRepository.java
@@ -4,9 +4,7 @@ import com.example.happy_fitness.entity.MailTemplate;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface MailTemplateRepository extends JpaRepository<MailTemplate, BigInteger> {
+public interface MailTemplateRepository extends JpaRepository<MailTemplate, Long> {
     MailTemplate findByCode(String code);
 }
