+++ src/main/java/com/example/happy_fitness/repository/OrderRepository.java
@@ -4,8 +4,6 @@ import com.example.happy_fitness.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface OrderRepository extends JpaRepository<Order, BigInteger> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
