+++ src/main/java/com/example/happy_fitness/repository/SupplierRepository.java
@@ -4,8 +4,6 @@ import com.example.happy_fitness.entity.Supplier;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface SupplierRepository extends JpaRepository<Supplier, BigInteger> {
+public interface SupplierRepository extends JpaRepository<Supplier, Long> {
 }
