+++ src/main/java/com/example/happy_fitness/repository/TicketRepository.java
@@ -4,10 +4,9 @@ import com.example.happy_fitness.entity.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Ticket, BigInteger> {
-    List<Ticket> findAllByFacility_Id(BigInteger facilityId);
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
+    List<Ticket> findAllByFacility_Id(Long facilityId);
 }
