+++ src/main/java/com/example/happy_fitness/repository/VoucherRepository.java
@@ -4,10 +4,9 @@ import com.example.happy_fitness.entity.Voucher;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface VoucherRepository extends JpaRepository<Voucher, BigInteger> {
+public interface VoucherRepository extends JpaRepository<Voucher, Long> {
     Optional<Voucher> findByCode(String code);
 }
