+++ src/main/java/com/example/happy_fitness/service/BlogService.java
@@ -1,6 +1,5 @@
 package com.example.happy_fitness.service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import com.example.happy_fitness.entity.Blog;
@@ -10,7 +9,7 @@ import com.example.happy_fitness.entity.Blog;
 public interface BlogService {
 	Blog saveBlog(Blog blog);
 	List<Blog> getAllBlog();
-	Blog getBlogById(BigInteger id);
-	Blog updateBlog(Blog blog, BigInteger id);
-	void deleteBlog(BigInteger id);
+	Blog getBlogById(Long id);
+	Blog updateBlog(Blog blog, Long id);
+	void deleteBlog(Long id);
 }
