+++ src/main/java/com/example/happy_fitness/service/CoachService.java
@@ -2,10 +2,9 @@ package com.example.happy_fitness.service;
 
 import com.example.happy_fitness.entity.User;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface CoachService {
 	List<User> getAllCoach();
-	User getCoachById(BigInteger id);
+	User getCoachById(Long id);
 }
