+++ src/main/java/com/example/happy_fitness/service/FacilityService.java
@@ -2,12 +2,11 @@ package com.example.happy_fitness.service;
 
 import com.example.happy_fitness.entity.Facility;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface FacilityService extends BaseService<Facility> {
 
     List<Facility> findAll();
 
-    Facility findById(BigInteger coachId);
+    Facility findById(Long coachId);
 }
