+++ src/main/java/com/example/happy_fitness/service/OrderService.java
@@ -5,13 +5,12 @@ import com.example.happy_fitness.dto.OrderDto;
 import com.example.happy_fitness.entity.Order;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface OrderService extends BaseService<Order> {
-    String order(List<BigInteger> carts, String voucherCode);
+    String order(List<Long> carts, String voucherCode);
 
     List<OrderDto> findOrders(UserDetails userDetails);
 
-    List<OrderDetailDto> findOrderDetail(BigInteger id, UserDetails userDetails);
+    List<OrderDetailDto> findOrderDetail(Long id, UserDetails userDetails);
 }
